backend/requirements.txt
fastapi==0.123.4
uvicorn[standard]==0.34.0
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic[email]==2.10.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator==2.1.1
twilio==9.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0
python-dotenv==1.0.1
